Fix a http put race condition in the client.
The race was:
When a put is done, both write and read callbacks are registered. If the
read callback before the write callback, then a dangling read error.
This was exhibited in an assertion failure in the gass proto interface,
because the send_done call was called twice (once from the write callback,
and once from the read response callback).