The previous commits corrupted main.  This check in hopes to correct the indroduced earlier.  Comments, and code where lost from the 2.3 release all of the missing items should now be back.
Changes were made to names associated with replication.  The word "master"
has been changed to "primary", "slave" is now "secondary", and "server info"
is "replica" info.

A bug was found in myproxy-info.  If the system was configured to use
replication, a user could not get info on certificates from a secondary.
This bug was corrected.

Modified Files:
 	Makefile.am README.building VERSION auth_pam.c gsi_socket.c
 	gsi_socket.h myproxy-server.config myproxy-test myproxy.c
 	myproxy.h myproxy_authorization.c myproxy_constants.h
 	myproxy_delegation.c myproxy_protocol.h myproxy_read_pass.c
 	myproxy_sasl_client.c myproxy_server.c myproxy_server.h
 	myproxy_server_config.c myproxy_store.c