Fixing a strange dormant bug that the NL control messaging code trips.
All messages are gss wrapped, and then base64 encoded.  When gss wrapped
the new message could be larger or smaller than the original, it all
depends on the magic of openssl.  Base64 messages are always bigger.
This code assumed that the Base64 message size would always be the
biggerst of the 3.  This is not true.  If gss compresses the original
message it is possible that the orginal message is larger than the
base64 encoded version of it.