Add Jenkinsfile to build
This patch adds the Jenkinsfile support using the helpers in the
gcs-build-scripts repo. Also, all metadata is refreshed, and tests
are adapted to work in a chroot without network connectivity.

All configure.ac scripts are updated to allow suffixes like ~rc1 to
support the automatic workflow to end up in the testing repo.